add doc, move findere libs to a separated folder, remove truth and BF fom findere, correct findere's query
