+++ tests/test_BF.cpp
@@ -5,10 +5,36 @@
 #include <string>
 
 #include "../src/libraries/evaluation/evaluation.hpp"
+#include "../src/libraries/findere/indexer.hpp"
+#include "../src/libraries/findere/querier.hpp"
 #include "../src/libraries/indexer/indexer.hpp"
 #include "../src/libraries/querier/querier.hpp"
-#include "../src/libraries/similarity/similarity.hpp"
-#include "../src/libraries/utils/utils.hpp"
+
+class bfAMQ : public customAMQ {
+   private:
+    bf::basic_bloom_filter* _bf;
+
+   public:
+    bfAMQ(bf::basic_bloom_filter* bf) : _bf(bf) {
+    }
+
+    bool contains(const std::string& x) const {
+        return _bf->lookup(x);
+    }
+};
+
+class truthAMQ : public customAMQ {
+   private:
+    robin_hood::unordered_set<std::string> _t;
+
+   public:
+    truthAMQ(const robin_hood::unordered_set<std::string>& t) : _t(t) {
+    }
+
+    bool contains(const std::string& x) const {
+        return _t.contains(x);
+    }
+};
 
 TEST(TestBF, TestFPR) {
     robin_hood::unordered_set<std::string> truthPlusK;
@@ -17,13 +43,13 @@ TEST(TestBF, TestFPR) {
     const int epsilon_percent = 5;
 
     std::vector<std::string> input_filenames = {"data/ecoli2.fasta", "data/ecoli3.fasta", "data/Listeria phage.fasta", "data/Penicillium chrysogenum.fasta"};
-    std::string querySeq = extractContentFromFasta("data/Salmonella enterica.fasta");
+    std::string query_filename = "data/Salmonella enterica.fasta";
 
     // create a truth and filter
-    const auto& [truth, filter, x, y] = indexFastas(input_filenames, numHashes, k, epsilon_percent);
+    const auto& [truth, filter, x, y] = indexBio(input_filenames, numHashes, k, epsilon_percent, 0);
 
-    std::vector<bool> truthQuery = truth::queryTruth(truth, querySeq, k);
-    std::vector<bool> responseQuery = noQTF::query(filter, querySeq, k);
+    std::vector<bool> truthQuery = truth::query_all(query_filename, truthAMQ(truth), k);
+    std::vector<bool> responseQuery = findere::query_all(query_filename, bfAMQ(filter), k, 0);
 
     const auto& [TP, TN, FP, FN] = findere_internal::getScore(truthQuery, responseQuery);
     double fpr = (double)(100 * FP) / (double)(FP + TN);
