--- tests/testutils.hpp
@@ -1,2 +1,0 @@
-#include <string>
-std::string randomSeq = "GATTATCAACTCCCCATGTTAAAGAGCGGTAGCGCCCGTGGTACTTGATGCTACATTTATTCTTTTAGGAAGACGGATACCTCAACCAAGTACAGGTCTTCAAACGAGCTCTGGGCCCCCCCAAGTTCCGCGTAGAACTTAGTGATCTAACCCGCACTAGCACTGGTGAGAAGGTTGACAACCTCTTCATTAGTTGTGTGTGTGCGTTATGTCATAGACGGTTTATAGGGTTGACCTCATGGGTCCCGACATATACTGGGAATCATATTCTTCGACCGGGGAGGTAAATACAACGGGTATCATCGTTATCATCCGGGTAGAGATGACTAAAAACGCTGCTAGGATAAGTATCATTTTATAGAGAAACTGGTACGAGGAAGATGCTTCGGGCGCGGTAGAACGGTTAGCTAATACCGCGAGCCCAGAACCAGAAAAAGTCTCTGCGCGTTGGTAGCCAAGCCATATAGCGTTTCCGGTAACGGGGTAAGTAGCGTGGCTTCTACGCCGACCGCTCACCTGCAACTAGCCTTAAGCAATGTACCTATGCTTGAATTCTTAGTCTTTAGGACCCGTGTCTGGAATCTTCAACGAGAAGTCCTCTGTCGAGGCCTGCAAAGGAATGTAAATGCTGATAGATCCTTAAAAAATAACGCGAACGTCCTGGTGGGACCCGCCCGAAGTCTTCATAGCCGCACAAGGAAAAGCAATTTGACCTATGATAGTACATAGAGGGCTGAGCTGGAGCGCGCACGAGGAGAGCCGTTCGTGATTAACGGAACCACTTGACGCTATCCCAGTGCCTCTAGCGAATGGCAACAGGGTTCGGACGCTCCACGAGCAGTTAAGGATAAAGAGTTAATGGAGGAAATGTGTAGCCAGTAGGCTTGTTCTGCATGATTGGTGACTTACCCGACGGCTGCTGTAATGGTGCTCAGCATTAATGGACAGAATCGCGCTCCACCCTGACAACCAATGGAGATCGGAGGCCGGCCCGCTGGTACTTGTGAATACTTAGCACCTCGGTATAGTACCGCTGTGTGCTGGCCTGGTGTACCCGTTGAGAGCAACATCTTCGAATAGATGCATGGCTTCAGAGCGGACGAGGGCGTTAAAAGACTCATGTGTCCTTGGAAAAAGGGAAACCGTAGCAGCGCAAAGAACTATTGCAGTCGTACTCAGGACGGATGGTATCAAAATGGGAGGCTTCCACACAGCCACCAAGGATTGAACTGGCAAGAAAGGTAGGAAGAAAGGGTTATATAACGCTCGACCGACAACAGCGAACTTAAATAATTGTGAGCATGATTTTACGTCGGGCGCTATTTCCCGTGTAAAGAGCGCGTCAAGATACCTAGCAGGGGGGCGAAGTTAACCAGTGAGCTGTTCACGATCGGGGTGGTTCAATGGAGAGGTGCATCTACTGGCGGTTAAGTATGCATGGCCCGTGGATGGTCCCCCAAACCTGCAAAACTAATAGTGTCAGCTTTACCAATTCACTATGAATTATTTTATGTTGTTCCGACTTACTCGTTCGATTACCAATGTTCGGCTTAAACACTCACGCCACCTAGTTAGGGATCTCTATTGATGATTCCTAAAGGCGAGTGCCAAGGTGTGGAGGAGATACCAGACCGATCACGTGGGTCGGTGATGCTCCGGCCACACAGGGTTCTCCATACTCCTGGAGGGTTCACACGTCTAATAACGACAGAGTCTCCAGGGCCTTGAACTTTAGAAGCTCGCGCTCTTTTGGTAGATGCAAGCTTATGCTACAAAGGGGGCCGACTGTTTTACGATACTCAAAGGAGGTGTATAGTCGGAGCGATCAGGGATCAGTGCAAACCTGGATGACGTTGATCGCTAGAGTGTCGAAAAGCCCCCAAGTCCAGAATTCCCAACTTTGACGTCCAGAATATGGCTTCCATCTTAGGTAATGGTGTGCAACGGCGTTTATTAGTGCTCCACAGACCCATGAGAAAATAATCCTGTACCTGTTTAAGCTGCAGAGATAAGGCCTCGGTCCTGGAGGCTGCGAATTCATGGTATGTCTATCTTGCGATAGCATCACTGGAAGGGTCAGGGATAATCACCGCATCCGGATATGAGTATCAGCTTCTGAGTGGTACTTAACTCTTTCCGTACATAACCCGTCAGATTAATACTGTCACTCCGTCTTGTACGTAATTCTTCGTTCTCGAGGGCTGCTTATGACAGGTTCGGGCCTGTCATCCGCGATTCTGTCAATGGCAACCTGGGGTAAACGGTTGGTCTATGTCGCCCCTCGCTCTGGTTCCATTGCGCCACGGACATGACATTTCCGACGGCAGCTTAAGGAAAGAGGATATCTTTCGCAGGGTGGAGGGCTCCGGCCTTGTCGCGTAACTGAATTGTTCCGGAGCTGTAGGCAGGGCAGACTTGATCGCTCACAAAGCTGCGCAACAAGTCGGATGCGTAGCACGTCAGTTCCATGATAGTACGTGAAGTCTTCAGACATGCGCATACTTAGCATTGACTAGGGCCGCACCCCTGAAGATAAATGCTGCCGCCCACAGTGTTACTGCGCAAGAGCGCATCCTGAGAGCGACTAGCGCAGTCTTGGATCATTTTACACGAAGAAATGCGGGTCTAGTCTATTCATACTTCAAGATGGGCAGGTAGCAGGGTGGCTCACTTTGTCAAGTAGGACCTGGCGCTGCCTGGACCATAGCCTGTTCCCGGCCCGGTTGGCAGTAACTATTTGCTCGCTAGCCGCTCCTGAACCACTAGCCCAGCGTGCCCACCCTCCGCATAGTGATAGCAGATATTAAAGAGCTCACGCTAAGCAGGAAAAGAATCTTTACCACGCGTAGTAGCAGGACTCTAGTGCATTGCTTGCTGATCATCGACAATTTAGTTGTGAATTGTGACATCAACCTATATCAGGTGCAGGTGGGCATTAGGTCAGCCCTCCTCGTCCACGTTACCACCTTCACCCGAAGGAATAGGCATGGTAAGGGCCAGGGATGTGGCCAGGTTGCAAAAGAAAGCCCAGAAGACGCAGACTTTCCACCCCGATCGCGTTATACAGCAAACAGTATTAACGCCTAGTCCTTCCCCTTGAATCGCTAGTTGTTGACAGGTAATAGTTTCCGGCTTATTGTCAACTTCACAAACTGTGCATAACAATGTCGTGATGCTTTAAGGGACTCCTACACAAATCCTAGACTGGATGTGCCTGGTGACCCGTCAGCCACCTTCTCATGAACCGGTGAGAAGCGGTTAAAGGCGAACTTCGTATAGCTCTTCCTCTAAGGCTGCGGACGTGTCAAGGGCCCGTGCCTAATGCGCGTGAATGGTCGTGAGGAGGGGATTCATAGCACGGATGAACGCGTTGAATGCAATGCGTAGACCCACATGGCTGTGATGAGGGGATGCATCGATAGAACGGGGGCTGTGAGCGCGCTTGTATAAGTACGCGTGGGGCATCCTGACCGCCACGCCGGACCAAAAAGGGTGGTGACTGGGAAGGCAAAAATCGGGGCGGTGTAGAGCTAAGGTTCTGGCTGCTATGACGAACTGCGGGCGTCTGGACCAGCGTTCATCAGGAGAACTATCCGGATCGAGCCGGGATCGGGTCGTTCCCTCCTTTGAGAAAACGACGGTGGATCGACGACGAACCCTTTGTCGAGAGGCTGCAGCCTCTTCGTTCTTCACTTTGTTGAGATGTATCTAGATCCATTTTTCGTTAAGGGTAAGATTGATGCCATAAACTAAAGCTCGGAGACCAGCCCACGACGGGCAGCACAGTTTCTCCCAGTAATCCTGGCTCACCCCGGATTGCTATAGCGCACCCTTAGACGGGCTACAAGGTATAAATGCAATGTTCCGTCAATGAATCGATTCCCTCGTTCCGAAGGTGAGGCCCACACATACTTTGCGAGTATTTTGGAAGAATGGGCGGCCTTGCCTGGCTGGTGCTCATCCTACGGCTAACGAAGCCGCAAAAAACGTTTGTGCATCTTTCTGCTAGATACAAGGAAAGTTTATGAGTGATACTAATTGATGATCTGACGTCCGGGCACATGTAGGTGTAGATAAACGTATAGTCAGACTCCAGCTATGCGGTTCAGCTCAGGTCAAAACACCGACAGCGCCCACTGCATAGTAGTCGGTTAGGCACGGTGACACCCAGGGATGGTAGTCCTGGACGGCGCCCTCCCCCTCCTCGGACCCAGATTATAGGTTAGTTGGGATATCTAGTCGTAAATTTCGGTACGGAACTTCTCTCGTGTCCAGATCCGAGAAACACTGACTGTGTCGAAACAAATACTTTCTGTAAACCATAACTTGGCGCGCAAGCACAGTGTGGTGCAATGATCGATGCGAAGTCGAAGACGTTAGTGACTCAAATCGGTGAATGGACTAGATCCAAGATATTTGGTGGGTCAGGCCTCACAATACATGGCGAGGATAATCTCCCATAGAGCTTAGATTTTAACATCCTAAGCGCGTCCACTTGATTGGTGAAGTGGTGCGTAGAGTTCTGTCTCAATATTACATCCGACGAAGTTGAGCTCTCCGGTATCGATGTCAGAGCAGCGGCGAGCCTATCTGCAATGAAATAACGACCGCAACTCCTCCGCCCAATTCAGCCAGCTAATAAGAGCGAGGTAGCCTGGTCATACGTATCGCGCCTCCATTCCTCATTACCAACTATTGAGCCATCTCTTTAAGTGACTCGCTAGCCGGGGTGTGCCAGGGGGCCTGGACGCCACTAATTCTATCCGAGTATTGCTAGATTGTTTAAAACCTCCCGTCGTCACAGTTTGCTAAATGCTCCGAACGAAATTCCGGGTAGGACGGTAGGGGGTTCAGAGGGTAGTATTGACAAGTGCGAGATGCTAGGTATTATTCTTAGTAGAATCCTGAAGTTTGATTCCTTCGGTCCTGGACGTAGCAAGGATGCGTGCCTGCCATACCTTAAAAGATTTTTAAACTGAATACCTACAGGACACTCAGGCCCCCACAGGATATAATTCCTGCCTAAAAAGCGTCCCTTCTATTCTGTATAAAGCCGATAAAATTGGCAGAATGCCTGCCTATTGAATGATTAGATTTCTTGATAGATCATTGCAATAAATTCTAAGACCACCGGGAGAGGACCGGCGCGGAAGTTGTTCTATGCTTTGTTCGAGTGTGGGTACTATGTAGGTTGAAGCCCGCTTGGTTGGATGCGTGACACGCGTTTAGTTCAATAGAAGTGGGAGTGCAACTCCCCTAGGTCGGATCCCTACTCTACAGTTATACACAACACAACCAACATTCACGCACGATCAGGGCTGTGTGCAATGGTAGTTGTCCAACAATTGCTATTGCTGCGCGATGAGCAACGGCTATTTAGCGCAAGAGCGACTCAAGGAACGAGTCATCATGACCATAACTTGCCAGGCCCCTTCTACACAGAATAAGCCGGCTAGGTGCTTAGTAACGGCAGTCGTTAACCTTCTTCGAGCACACTATTGATTCCTTACTCCCATGGCTCTGGAAGCAACCAGCTCCACGTGGCACCGGGTTCGCAGGGGAGGTTTGATCACAACCCTTACGAGATCCCTCCCATGCGACGCAGTCACGTGCAGAACGTGTATAAAGATCATCCATTGTAGCGTAAAGAGCGGAGTGTGTTAAGAATCGATTTGTATCGAGATTCCCTATACAAAAAGGGGTGGTCAAACTCATGTGACCGAATGGAAGGGTAGCCGCAATTGTGCTCTGTCTTTGACATACGCTGCCTGGGGGTATCGTGTCATTCTGCGCCTCGCAGGTCGTTCCCTTCTGACGACTAGTACTGATTATTTTCAAGCCAGACCGTAGACTAGAGTCCAATCAGTCAGCCTTTTAATGTCATGTTCCGCCCTCTGACGAGTGCATTTGTATATCCAAGACAATCTTACCAACCCGATTTTGATGAACGAGAGGGTTTGATCGACAGATGTGAAATGTTACATAAGTCCCGCCAACGCCACCCGACCTACAAGATGCCACGGTAAGCATCACCCCAATATAGTCGGAAGCATATCGTACTACGTGTTGACTAACGGACCACCGCCTTCGACCATGGATTCACAGCCTCAGCTAGGGCTCAGAATAGGGAAACAGACGTAATTTCAGTTCCACCCTTGCTGTTATTCAGCGCGGAGAAATTGTCAGACAACGACACGGCATCTCAGCTACTCATTATGTAGCAAGATATCTTCCGACATGGGGTATTTGCCAGGAAATAATTTTCCCGAGACATCTAGTAGCCCGCTTAGCCCTCGCGTCAAGTCAGGGGGTCCTTTGCCACATGGTTCTACAATATACGGTAGGTCCACGCTTATGAGCCTAATCCTACTCGTGTTGGTATAGGTGAGGCAGGAAAAGTGTCTTCCTTTGTCTCACGGGGTTTAACATGCGCACCTTTAAGGTACCGCGTTGAGCATGCGGTTCATCCGCGGCAGCGAGCTGCTAGGAGTTCAGCAGATCGAGCGTGTCTTTGTCTCCCCACATCCGTAAGTTCCTGAGCGCCGCGTTATGGGGCTCGGAGACGGACCTTGTCACAACAGTTGGAAGCCGGTATATTGTGTTAGGTCTGAGTGGTATTCTGACAAAGGAAATGCACTGTGAGTAAGCCCCTCTCAGGCGGACACCTGCGATGTGCTTAGTCCCCAACGATGCCCCTCAACGAGAACGTGCTTTAGGTGACCAACCGCCTGCAGTTAAATATGACGGTTCGACACAGGTAGGTTTCCGCCACCACTGCATATCCCTATACTGGACCGGTGACCCGGCCGTCCTGTTAACACTATCTCCGAACAAGTTTCGAGAGTCTGGACCCATCCTTTCGCGTACGGGATATCGCGTGACTCTCGGACGGATCACGTTGCATGATGTCTAGAGTTGCTTAGGATTCCCAGCCTTCTCAATCTTCTACCTGACTACGCCCGCAAGCCGACTACTTAACCTGTATTTGCTCCCCAGAGGCTGCAGATAGTAAAAGATGTGTGGAACCGCAACAAGGACGAAGCCGATCACGCTTCGACCTCTTGCCACATTGCGTTTCTTCAGCTGATATATGAACTGTAGTAGTCTTCATCTCGGGGATACTAGCCCTATAACCTGGAGCGCAGTTGAAGGTGCAATATCCTGCATGTAGGCGACCGCGTCAGATTCCCCCGATCCAATTGGAGCAGGAACAGTAAGTCCGACGATCCGCTACGTTTCATGCGTGCCGATATGTACACGTCCCCGTTTTTGATCCTGGCACATGTGCTTCGGTAGGGACCAGGGAGGTAGACCTGCAGTCGGGTTCGTTAGCTCGCAAAGCATTTGGGCTTGTACTATCTACAGCTGGGGGGGGATATCAGATACATAACGACCGTATGGGCTTGCGAGAACCATGGCTTCTCGAAATCGACCTCAAATCCTTGTCGGCTCTGTTCAATGGTGGTAACCGCAGCACTGCTTTAGGTCATCGGTGTTAATATGCTCTGTCTTTGCTGTCATTAGGCGAGTGTTACGCCTAATCTCAATCTGAAACTACGAAAAATCAGAGTCAACATGTGCCAATACCCGGACTGTTATTGTCTAGGACTGGTCAGACTCTCCCCGCCTTACAATCATCAACAAGATGTAGCACCGTTTCTGCAAATCCTACCGGGCTGTGGTTACACCGCGGATCGTGTCTACTTTGACCGAAATGGTCTGACATTCAGGATAATTGGCGCGCAAAGTACCCGCTCAGGTTATATGCTCGATAGGGCTCTGACGTCAGAATGGCTATCCGTACGAGAGCGATATGTTGTAACACCGAACTCTGACCTACCACCTTCTATTGAGCCAATCATGCGAGACTTGGAAATCAGTGTCTTAGAGGATAAATAATCATGTATCAAAGTGTTCATCCTATTATGACCCTGTTGCTGATGGATACGAGAGGGTTTATTATTTTCAGGGTGCATGCCACCGTCTACCGAGCCCCCACAATCCGGTGACAGCGCAACGTCGTGGAGATGACTTGCTTTCCCTCAGGACTGAGCATGTTATACCATTGGGGCCACCGTAGTGCAAACATAGCAGTAAAGGAATTTAATCTTGTCCGGTCACACGTGCGGAGATGCGTAACTGAGGGCCCGAAGCCGACTTCCCGGCTAGGAGCCTCTCTATTCGATTCCCATTCCAAAATCCGCCGGCTTCCGCTGTCCCCTTAATCCGCACTCAGAATACTCTCGAGCTATTCGGGCCCGATAAGACGTCGGGTATGCTTGCAAGACGTGGTAGCAAACACCGTTTCTATGGGCGTGTGGAGAGTCACGTCTGGTTCCATGATGGTATCCCGCTATCTCTTTTCAACTCCCGGCGTTCAAAGCCGATTACCGTCCTAGATTCCTCTGTGAGGATTAAAGGTGAGCCATTCCATCGTGCTCCGTGTGATGACGAAGGAAATGACGTACTAATCGAGCTGGATTAGGCGCGTACCTAATTCCCGAACTTGTGAATGTTCGATTTGGATCTGAGCATCTTCACCCCTGTCATTCGCTCCACCTCCTATTTCTTGTAAATGAATAAGAGACCTGGATGAAGTCATGTCTCCGGCTTATCCTTGCAATCAGCTGTCTGTAACTGGCACGAATCCTATTGTGTAATCTCTGCTTGACATGGGTGTGCGTTAGTCGAAACTATCACTATCTCTCTTATCCCACGGGCTCGCTTAACGTCACTATCCAGTAGGGAACAGAATGGTTAATTTCAAGAATGGCAGCCATGTGGTCCGTCCACGCCTGCTGAAGTACGCCGCAGCAATCATCCGACCGGATCTCCCTGCATTCGAGCTATTCACAAAAATAACAGAGAACGATATTTTTATGCCCAAAAAACGCCCGACGATCGCCACGCGTGATCACACTAGTAGCACCGAAATAACGAGGGCGAACCCTCAACATCCGTATACAATTCTGTGCCGGTGAGGCAAGATGCCATGTTATAGTTATTAACGAAAAGACGATGCTCTGCGTTCGGCATCGTAAACGTAGGAACGAGAGGGCCCATACCGGTATCCACCTGGTTGAGCCAGGAAAAAAGATCTCAGTCCTGTCAATAGAGTATTTCATACGTCAAGAAAAAAGCTGTGAAAGTCTCACTGGGTTGGAGTGCTGTCCTTATTAGACACTGGAGGGCCGGGTCGGCTGACCGCTAAGCCACCCAGTCGATCCCGTTCATCCCACCCCGCAGAGCGTCCGGCTATATGGTAGTATCCCATCGTTGCACGTCTGCAAGTACTGGAGCATCCCGAACTTCATAGTAGGTTGAGCTTCAAATATGCCGCCTCTACGTTTTTGGGCATTGTACCGCAAGGTCTATAACTTTGGACGCGACAGTGCCTCGCGGTTGCAAATCCGGTACTTTGTAATATGGACGTCACGATGATTGAGGATGGGTCATATTACCCTTGTCGTGAGACCGCTAACGAGCGCACACCCCAGAACATCCAATTCGCATAATCTCCTACTTCAAGCCATACCAGTCTCTATGTCTATAGGGCTGGAGCCTTAGGTCATGGCACAGTATCAGGGACTATGGTGTTGGAATGTTCGTACTGTATTTATAACTAGCAGTCAACGATCCATCTCGAGGAGGGACGATGTGTAACTTAGACTTCTATAGTCGTGAATCATACTGGTGAGTCTGGCTCAAAGATTTAACCTGAGAGTCTTTAGGCCTCTGAACGTGGAATGAAGTATTTGAAATAGAATTCGATCTTAGAATGATCCCTCACGCGCCCTGTGTGCAGAGTAGGTACCCGCATGTCGTGCGGCATGGCCGCGGAGGTGTCCGCGACTTCCGGTGTGGCAACGGTTGCCTTGTGAGGACGCATCTTAGAGTTGCCTCAGTACCACCCTCCGAGATTGTAGGCATAAGGCTGTCGAATCTACGAAGCCAGCAGCTAGCACCGGAAGAAATAGTTATGTAGGGCACGGGGTTCTATTAGCCCCTGGCTCTAATCAACGAAACGAAGCCGTTTCTAGTGGTTAAGCCAAAAATTAACGCTGACTCTAGGCATGCCCTGAACTTGCCACTACGTTCGGCTTGGAATCGAACCCAACGAAGAGTAAGGTACGTAATGCGTCTCGTGATGACTTAGCTGGCCGCGACGCAATCAGGCCTCGTGGCAGGCGTCGATCGGGTGAACACTTTTTCGCTAGGATACCTGAAAAAAATAGTCGACTCACGGTGTGGACGGTTTTTTATCCAGTGGCGGCGCGTATGCACGCCACCTAATTATCATTGCGATTGGGAATATGAGTGTCTTGGTTCGCACTTCACTAAGGGCTACTACATTCGCAGAGGAATACAGAATACACGATGACCGTACTGCTCTTTGTAAACTTCGAGACATACCTTCTGGCTAGCCGATCCAGCTTCGGGATATTTGGCGAGTCAATCCAACACGGGTAGAAGAAGGGGATGATAAGGTATTCGGCCACTGTTAGGTTACTGCAAATCGCGGGGCTCCCGCTCTTGAAAGTACAAAATTACCTAGCATCCGAGTCCATGAGAACTAGCCTTCCGTTCGTAATAGTGGGCGCACTCAGGTGGATCCGGCTTGCATCGAGATCAAACATCCTTCGCAATCGGCCACGGCATGGGCTAAGTCGTTATCCAACCCATAACGCTGCAACAATCGCCCGGTCACCCCACACTAACGTTCCATACGCGTCTATGACGAACCTATGCACCTAGATCCTGTGCGGACGTGATTGTTGCCCTTGAAGTTGATATTCGCAACGTTGCGTTCAACCAACAATCGGCTTGCTTAACAGTGATATGATAACTATTTTCACGGCTGCAAAGTTACTTTTCATACTGGATGAACCACAACTAATCTCTCCTGTCACGGTCGCCATGATTTCAATAGCACGATCCCTGTTAGCACGGAGTGCCTTTATACGAATCATACAACAGATTCGACTAAACCGGGACAGTTTGCCATCTAAGAAAAGATTTAACAAGCTCGGTAACCAGACACTAAGTATTTGTATCGCAAGGTGATTAGAAATGGTTTTCTCACGGTTGGCCCCACACTTTATGCTTCGAACTCCTAAGTAAACTTTGCAATACCAGACATCAGTGATCCTGCTATATTGTGGCTGGCGGGGCAAATGGCGGACATGGACCCGTAAGCGCCCAACGATTTGCCCATAAGTCCCGGGGTAGCACGGATGGGGTCTATCAATCACTGTGTAGTCGTGCTGCGACGCCGGGGAGAGTTCATTAGTGTGAGCTCTCAGGCAATAGCATAATCGCACGACCACCCATAAGAAGTATTTGCCTCCTCGTTTAGCGGCTAATTTTGAGTCAGGAGGTATCATCTTGGCTTTGTCAAGCCTTTCCTGTCCATGAAGCATAGTAATCTGAGTGTACAAAAAGCCATTTGCTGCATAGGGTAGTGAATAAGCCAACATGGGGACTTACCTTACCGTGGGATTCCATTCAAGAAATTCTGGGGTGTTGACCCACTTTACGCCAACTGTTGGCGGATAAGGCGCTACGTGACTGCCCACGTGCTCGGCACGACTATTATTCGATCGGAGGGAGCTTAGTGGGCCAGCGCCCAGCTTGGTGAAACTAACAACAACCAAGAGCACGACACTCTGTCTTTATGGTCTTCAGAAGAGGCATTCCAAACCGAGTGTAGAACGTTACGGACGACCGTTGGTACCTCAGTAACCTACGTCACTGTCAGAAGCCCTGGTTCTTGTTTGCCCGCTTCCTTAGGTAGTCAGGGAGTGGCGTTATTTCTTGAGCCAGCAAACCCTTTGGGGGTCCGAGCCTCCAATGCCTAACGTCCCCTCTTGACCAGCTGACGGCCTGGAGGACAGGGAACAATTGGCGTCATTGACGAGAGATGTAGCTCGGTGATTGAGCATTGGACTGCCAGGCTATTGTTAGAGCTCCTAGAGTATTCGCGCATTCGTAAGAACCTTCTGATATATTCGGGGTGTTTATTACTGAAGGTGGGGTCGCTTACCGCTTGTAACATTGCTCTTGTGGTGTGGATCGGGGTCCGCGGCTTGGTTGGTTAACCTGATTGAGTCATACCGCATTGGTCAATACTGTCCCACGTATACCGTGGCACTAGAATTTCTATATTTTACACTGAGGAAGAACAAGTTAGCGTGCCGAACGGGGCATTTTGCGGCCCCACCAGCCGGGCACATACAGTGATAGGACACATCTGGACTCCTCAATGAGTATTGTATCAAGTGCGGTGGCACCAGCAAAGGCAATAGAGGTCATCTGACAGTTGACGTAAGGAACGCGTTCCATACTAAAGTCCTTTTCGAAGTGTAGCTCACAGTGTGGGAGCGCTTATCCCTGACCGAACCGTCGGCGCTACGTAGAAATCCTAGTGTCCAGGCAGCACGTTCCAAGTTCGCCTGCAACCCTTAGACGCGATGCTGGTTGTAGATAAATAGTCATGTACCTGACGGCCAGAGCATGGTCTACTCACCTATAGAGAACTTAAGTGCGAGAAGAATGCTAAAGTCGACAGCTTCAAGAGTGCAGTAGAGAAAGAGATCACCGACGTTGTTCGTTTTCTGCGAAGTCTGGCCCTTCGCCCAGCGCACATGCCTTCGTAGAAAAATCAACGTGGACTCGCTTTTTAACTGTTAGTTTTATGCGCGCGAATTGGTTTAGGACGGTCGTGTGTGCGTACAGGTCGCGGTGGTTGTTCCCTGACGCAGATTCCGCTCCAAGTGATGGACCGTGATCCGGCTGGCTATAATTCGTTCATCGGCTGTATCCATAAATCAGCAACGGAAGGAGCGGATTACTTAATGTACCTTTTCTTACCTGTTCCAATTAGCTACAACATCATTACACATCCGTTTTGGATACAAACCTAAGCTGATGACCGCTCCGGGGGAGCTCCACGTGTTGGGTAATTAGAGACTCTCTCCAATGATTCCTATCTCCTCCAGCCTATAGGACCAACTGTACGTTCATACCTAACTAGTGGGCTAAGACTTCTGGGGACGCCACCGGAGCGCGAGTGATCGTAGACAGGCTCGCAGGCGATGATGCTAGTCTTCTAATGCTTGTTAATTTGGCCGACTCTAGGCTTACTTCGCTCTCCAGATGTATTACTAGTGCAGGGCCATGTTCTTCGAGCTTACAACGTGGAGAAGCCCGCCCCAAGACAGTCCACGTCAGACCCACTCGATCTCGGGGGTGAGGAGGACCGTTAATTTTAATCGGGGTGATTCTCCACCTCTTAGTGACTAGGCAACCGCAGATAGCCCCCTAACGTGAGCCTCTATGACTGCCGTTTAATTTTCGGCCTTAGTTGAGAGTAGCTTTTCCCGTTAAGAAAGACTCGATCGACTATCCGTACTAATCTTTCAAGCGGCCGGTTCATTTCCTCACCTGGCACTAGCCCAGTACCCAGGTGTTCGTGTTTATTTGTTGGGGCCCAAAGGCCTTTCGCATACCATATTAGAACACAGGTGCAACCCTCTGACTACCCGAGTCCTCGGAATTCACAACATTGGGTCAACCTAATTGCCAATGGCCGAATCTGCGCTCCCCTGTCGCTATTGAGGCGATGCGCACTATCCGCTACTGATACCTCACGGCTGGATGAGCGACGAACATAGCACAAGCAAGTTCCTTACGGGTGTTAAGCCGGGACGAAAATCATAATAATTCCTTACGCGGAGCCGCGACATTAGGCATTAACATAATCCAGAGGCTAAACGGATCCACACCCACCTTCATGCATCTAGTAAATAGTTGCGTAATATACGACCCACAGACCCGACCCACAGAGGCGTTTTTGGGGTTTTCAAGTACGCCGACCCCCCTCTTCGCAGGGTGCCGGAATTTGTTGTACCCACCCCCGCTTCCTTGCTAATTCTCTGATCGGAATGATTACTCTTGCTCGGTAGAATCACATCCGCTCGGCCCTGGCAAGGTCTGACACAGCCGTACGTTGACGTCAGAATAACTAGATGATCCGTTCTCCAGACAATATATGCTATCTTAAACCTAACCTCCCCTGGGTTGCGAATCAGGACCACTGGAAGCCGGCCGGCAATCTGCTGAATGGTTCGTCTATACAACACTGATCTTTTCCTTGGCAATCGGCCGAACTTGGGCACGTGACGACCTCCCCCCTGATGGTTTTTTTTCCGATAGCCGGCCGCCACAGCTCAATCATCAGTGTGCATGAGGAACCACTGAACTAGCGTCGTTGGTGGGAAAGAGAAAAAAGCCGTTGTCTCAACTCCGCCATTATCCTTCGGGCGGTCGCTGCGTTCCCCTAGAAACTTCCAGAGCAGACACACCAACCTGAGACTTAGATAAGGAACAGGAGTATTAGACAGCTAGGCTTACGACCTAATCTCACGGCCATCACCCATCGTAACCGGACCGAGGTTGTGAAGCGGCCTTCCGTGCGTGAATACACATCGGAGAAAAGAATAATCTTGAATAACCCCAAGGTGTCGCCGTACGGAGCCCTAACAGGTTTAGGATGGCATCTACCGGGCGACGTCGCACAACTCTTATGGAGTTGAAGAAGGGACGATAATCCCCGCGCTTATAGACGGGTCTAATCTCGACAGCTATCCCACTTATTAGCCTCGTCCCAGCAATAGTATTCACTATAGTCAAGAGTTCGAGTATAATCCTATTAACCCTATGGCGGCGGACTTCGCCACTTGCGGGAAGGGGGTGTGTAACTACGAGTAAGGTTACGGAAGAAAGTGTAACAAGTTGATCGGGAGAAGGTGCGCAAAACAACAGAAAACTCTAGCGATGGGACGATTGAGCCGTAGTCTAGGGTAGTGTTTGCAGAAACCGATAATACCGTCGTTAAACATTCGGTTGAGAGCAGCCCAAGCATGAGTAAGTGGTGCCCTGCAGAAACGGCTGTGTGAAGGACTACTCGTGCTTTTAGGAGCACTATAATTCTGATATCAGCCCGGGGTCGTTGGATGATATCCCTAGCCGTGGAATGTGCAGGACATAATTAGCAGTAACTAGCGCGCTTCACCGATACCCGCATCCGGTGATTGGAGTCAGCGGCTTAACCTTGTCTAACCCCCGGACAGAATTAAGGACGCATCATTGCATTCCTACGATTCCTCGACGGAAGCTTGAGAGAATAAAAGTGAAACAGTTCCTCTACGGGCCGGAGCGACCTGGGTCGGTAAAAACTTGTCTTGGCGAGTTACTCCGGGCTGAAGGAGATGATCTTTGCTCCCCCCGAGGTTCCTCGGCCACATCCGCGCCTGTTTCATCCCCTCAATGCGCACGGAACTACTATAGGATCGGTCGTCCAAGAGCTACGTGTACTTCTGACAACGCGCGAAGATAGACAAAGACGCTGCTCTTTTTATCATCTCTCTTCCATCAAATCCGCCTGGGATAGTCACTGGCAAAAACGCAATGATCGCGTTAGTAACTACAGATTCCGACGCGTTGTAAGTGGGTTAACTTTCAAATCTAAGGACAGCATGCGTCAAGTTCTCCAACAAGGGGGACTGAGAGTTAGCTTCCAGAATGCCTTGTAGGCAATCAAATGGTACTGTGATAAATATCGACAACAGTTGGACATACCTGTACCAAGACCGCAGCAGTAGACTTTACCATCAATCTTTTAATAGGAATTAGGCACATACTTCAGCATTCCAAGATAGAGAATGGTGCCCAGAAGCATCTAAAGGAGGCCAAGTACTCTCTTTCTGAGTCCTGCAACATCTAAAATAGTTTGCTGAGCATGCGCCAATACTGGGCGAACAAATGATTTCTGTGCGTTGGAGCCGATATGCGCAAACTAACCGGCACCTAGGGCACTCGTATAGTCCGTTCAAGGAACAGGGTGGTACAGTGCTGGGGAGATAATACCCGCCGGTCGCACGGGTTCCAATTGAAAGTCTGAAACTACTCGATCTAGTCCTAGGCAAACCCCTAACGAGCTTCAGTGCGTCGCTGTCATGGTCGGCCGATGAGCTAGGACCATGGTCGAGGGAGCTTAATTCACATGATTGCAATGATTTCACGCTAATGATTAGTAACGGGATTGACCTCTTCAAACCCGCGGTGGTCCAGCAGTCTACCTTTTGCTTATTACAAACGATATCATCCCAACCGCACAGGTACTCCGTACTAAGATGTTCCTTGTTTTAGAAAGGAAGCTAAGGGTAGACTTATGAGGGTGCATCTCGGCATATGCAACCTAGATCGCTACTTGTTCCACAAACGTCTGAGGCTTACCGGCCCATCGTGCCCGATACGAAGGTAATTTGCGCCATGGGATCCTTGAGTCAATCTACTACGCGGGACAAACGCCCTGGCGTGTAGTGGACAGTACTCAACTAATGAGGGTTTGGCTTGCTACCGCTACCCGCACGCTGTAGACCTCTACTCGCGCGCAACGACTCTGGAGAAACTCTATACAAGACCGGGCGACGTGTTTTCGCCTAGATCAAACGGACAGGCGCCGAGTTCCTGGCCCCATTGGTGTTATCGAAGGGTACTCATCCTGTGCCGGATTTCGTCGCGTCGGATATCCTACAGCAGATTTAGTCCGAGTTTCGCGATACATAAGTAATCACAATCGCGCCTAAATCATCCTAGCGCTTGTCAATCGCCACACCTTACCAAACTCCTTAAAGTTGCATCCCGTTCCGTATTTTAGGTGTTACAACTATAAACCGAAAGCAGGCTTTCGGATCTGTTATCATACGCGAAACTTTAAGATCTAGCCGGTCGATCCATCCGACTAACGTTGCGGAGGACGGTGGGAAATTCTAGGAAGCGAAAGGGTGCCTAGTGGGCCAAAGGTCGGCCGAGCTAAGAGCATCCAACACACCTCCCATGCTAGGATCAATCAGCAATCTAATCCCATCCGATACGAAACGGCGACAGCCAGCCCCGTTTTTCCCTGTTATCTGGTCGCGAGGTACGAGACTGTGCAAGTCGACTTCACCGGTGTACAAGAGACAACAATCAGTCCATTAATTGGTTTTCCGCGACACCCTATGAGGTCGGTTCTGCATGCGGCTCGAACTGGTGGTAGGTCTTCTCCTAAACCCTACCTACGGCGGCGGCCCCGGGGCCGCACGGGAAAGCGCGGGCCGTAGGCAATCATAAGGACTAGACGAGCAGATTAACTTTTCCCGTTTTGGTTGGAACAACAGTTTATATAGTCTGTCCTTCTGCTAGCGTCTGTCGTACAGGCGTGTGGCATGCTTAATCTTAACCATTTTACGAGAGTCAGATATGTTCCTCAGGACCGTAAGGTAGGTCACCTCCTCTTGTGACGGCTGTCTCGAACGTACCGCGTCCGTTGAGGGATATCTCCCGAAGTCGACGATTGTTACTCTGACGCCCGCCGCCGTTGCTGTAACCACAATTGCATATGGAATGATGTTCCATTGAACTAACGGTGTGCGAGGTCTTAGTTTGCACGAAATCCTTCGGGTAGATGGACGGAACAGTTACAGCGTGAGCGGACATCAACGCCAACCGTAAATTGCTAAACTTCTCGGGCCGGGCACGCCTGCCGGGACCCGTTATCCCATGGTCCTGCACAGATATACGGGCAGCCGACTCAAAAGAGCCGCTGGGGGTTTTTCATCAACCGAGAGATTCCGTCTGCCAAGCATAAGCAGACAGGTACAGGTGCCGCCTCGCGACAACTTACCATCTCTGACTTTTACCATAATCATGACGCGCGATCGGCCGGCCAACGCACAAAAGGTAGACCATTCGGCATACACAGTACAAGAAGTATGGCTATACTTTAACTAAGCGTGGCCATTATTCAAGTCAATTTAACACAATAACAAAAATCTCGCATATTCAACTCGCAGTAACCTGTTTATCTACAGGTACGGCCTGTTTGCGTGATGAACAGCAACAATGGGCCCATGCTAAAAGGTTTGTTTGGCTACACCTTATGGACACCATGGCTAAGCAATTCTAGTAAGAGTCGCTTAAGTGAAAAGTCCAGATACTCATGTCGCACAACCTGCAATAACCTGTGTGGCTCACCCATTTGCATGACGTGATCGCAGTAAACTCTCAAGGAGGTTTAGATGCGCCTTCCGTTCTTTGTGGAAGCATTGACGCGGCAAGGTTAGAGTGAGAGTCATCCGGCCCTCTCTCTCGCGCGATTCCTTTTAGGGCGGGACATACGGATATTAGTTCTGCCCGTATGCTTAACCTCGTTTTAGCTGTTTACAGCGACTAGCCTTGCGTGACTTCATGATTCCGACTGTCCGATAGTCGTTTATGCGGAGTAACTCTTAAAATTGTCGCCGATTAGAATCAGCTGAGGCTACGTCGTGTCAAAACTGTCAGCGGAAGCTGGCTATAGGCGTGAGCACATCCGTTTGCAATGCAATATTGCATAAGAATGGAGGTTCCCGCTGCTGTATTACGTGATAAAGTTTCTAGGTCTCCGCCGGCCATATCAGTGGCTGCTTAGGTGGTCGGGATTCGGGGTACCGCCGGGGCCAATTCTAATATACGTAGGTAAGCCGCTGAGCAGAGCTCTACGCCGGTGGAGTTCAATCAACGGAAAAGTACCAGTACAATATCAAGCGTAGGCGATTCTTGGTTTGTTGAGAGCAAAATTCCTCCTATCCGAGAGGCCACACGCAACCGCCTAAACCCCCATGGCAATGTATGTCCAACGAAAGCTCACCTTGTGTATATCAGTGGTCAATGGAACAAAGTGCAGTTCATCTGCACGACTCGGCGGCGGGGATTCTGACCTCTTTTATTGACGTTGGACCTTATCATTGACTGGTCCAAAGGCGATCGCTCGTGATATATGAAAAGGCCAGGGTCAAATAGCTGTTCCGTCGCCGCTGAGCAATTACCAGTATACTCCGTAGCGAGAAGAATCAAACCTGGTGCTCTCCTCCATAAAGCCCGTGCAGCTGCGTGAATGTGCCACCTGGTGGTGGTACGTTCTTCCAGGATCACCCTTGGTAATCGTGGTGACTCTGAAACAGATGTGTCGTAATGCTTCCCAACTACAAGCTTAGCTCATCTCACGTGGGAGTGTAACGCAGTGTGGGTGTTATAAGACATATGCCATAAGGACGCGCATTCGTCTTGCTTAACTAGCTTGAATCTCTTTGGCGCTCGCTCGGGTATAAAACCGCCTTGGTGATATTTTCGAGGGTGGGTAAGCCTAGAAACGCATATAGCGGTCGAGCTTCAATCCCCGTATGAGAGTGCTCATACTCGGGGGGTTGATGTAAACCATCCCCAACTGGACCGTGTATTACATTGTACCGACCTTGCGGTATACTTTTCAGGTCGTCAGGCTGAGTAAGTGCCAGTCAGCGGAACCCCGCATGCAAATACGAGCGTAAATGTCGCTAGGTCATATAAAAGGTCCTCCCACCCCAAATGGTCACGGCCCCTCCAACCCGAGGTGCTATCTCCCGTGCACTCCTTCGAGGTTCCAACGCTTGATGTACACTTTATAAACTACTAGGCTCTACCATTGACGTCTTATTTACTGTACACACACCCCTCTATCCAGTCAGAGCCAACCTTCGGCGTCCTCCTATCTTGGAGTTGCTTTTTGTCCGTTTTTGCCAGCACCAGCGTGGGAATCAAGATGAGTCATTGTGATACTCTGTGTGGTCATGTGGCTACTAAAGGCACGGTCCTTCACGCTATTGGCAGTGACCCGCCTCTGCTGTTCCAGACATAGTATGAACATCGCTGTATTGAGCCGCTGACAAGCTAGCTATCTCAGCACAGCTAATATGATTAAAGGGACAATCACTCCAGGTTTTTGAGGAAACGAGCATCTCTCCTTGAAGTGTCCCGGCCAAACGGTCCTAAAACCAGCGGTCCGACTTTGAGCGCGCAGCGAGTTGTGTTGAGGATTCGTGCTCTCTTGGTTCCGGTAGGAATCCTTCTTCGGAAACATAAGATCCTTTGCGTGCTAAACCAAGGGTTGGTATCGTGCACGCGCGCTGCCGATTGACGGAAGAAGGCACCAAGAGAATGACCCCAAAATATGTGCGGAGGAGAAGGCATGGGCTAGCGGTTTCGCTTTTTCTCTTCGCGAGCGTATACGTGATACGCCATTGCGCTTAACCTTCATTACAGGGCGAACTTTACTCCGGTCGCCAGCGGAGGCTAATCTGGACTTCTAAATCTCGTTTGTCAACGCGCTTCATACCCGCTACGTCCTGGGTTATATCTAATAGCGATAGACAAGTAACTCTGATTGGGGGGTTGCGGGCTTGTACATCCTATTGAGGTAGAATTATCTCAGCCTAAACGCAGGTGGCCCACTAACCGGCGAATAACGGTTGTGTAGTGTCGGCCACACCGGCCACAATCCAAGTCATGCGTACTTAGAATAGTAGAGTGACTGTTGCCGCGTTTTCAATAGGCCCATAATGGTACCACTATCGCGTCCCGCACTATTCCATTCATCTACGCTTCTATGGACGCCAGGAGTGATGAAGCTCATTCTGACACCGAGTGCCCACGCAATCAGAGTTGTGTAGTGGCCGATTGGCCTACCAATCACTGTGCTTCGCTATAGTGTCCCCCATAATCTAAGGGGGGATCCTTTAATCGGTCTCCTTCCGTCGCAATGGCGGCTGTTGCTAGACGGTATGGAGTAAACGGCTGCTACAACACATCGGAATGACGTTGTAGTTGGAGTTCTAGCCGGCATCCGTCAGAGTAAGCCACCAGTCTAGGCTGTTTGTATAATACCTTTATATTAGTATCATAACGCATTCAACAGCGCTCAGACGTCGGACAGTTCTACCTCAAACAAATCTCCCGAACGACCTCGTCTTCAGTGAGAACTTACTCTTTATCCACGCTCGTCGGAGCCGACCTGTTTCGAACAGGCGGGCGGTCTCGTCTGCGATTCAACATCCAACTCGACGGGCCGACAGCCTGTTGCGAGCTCTAAAGCCAGGCATCAGTGAGCAAGCTTGAACTACGCGTTGGCCAATTGACCTCAGAGGGCAGTCCCTTAGCGGACGTAGCCCAGCAACAAGACATCAATAGAAATTATGAGTTAATAATGGGCGCCGATCGAAGATGCATGTCGTGAATAACGAAAAGGAATCTATAGCCCCCAGTGGCTATCCAGACAGGCGGATTCCGGACCCTAAACACAAGTAAAAGTGGCGAACTTCTGAGGCATTAGACCGCGTATCAGTCAGGTCGACTTGGGGCTATTTTTCATCGTCAGCAGCTCAGGCTTCTTAAGGTGCTTTGTCCTTGAGCACAATCTTATATCTGCTGTCCAGTCCCATGGCTCCGAGAAACCGCGATCGATACGGGGAGAGACGCTTCGACGCATATTGACCCCGTCCCGGACATATATCCTAGGGCCACCGCCTGGCGCAGTTTAGCCAATGCTGCTCTTCTGTGAGCTAACTCTATGTATGCCTTAGACTCGAACCCTTATTTATACGACCTAGGACCAATCGAGTTGGGTGCGTTCTGATGTGGTTGGAAATATCTGTTCAGAGCCAGGAACAGGAGGACACTGGACGGCCGCTTCAGGCACCACATGTTTTGGTACGCAGGCACAGGAGTAACGGGCTTCGGATCGAAGGCACGGGTCCCTGTAGTGGCGGTGTGGCCAATTTTCAGGCGCCCGTTTGGGTGGAAGCAGATCCTACTTGAACTGCGGTCAAATGAACTACATCATCCGATCGGCATGTACTGGCGACTTCTTTTATTGATTCTGCCGATGCCCGAAGCAACACCCACTCCCAGGGTGGTTCCGCACGGGAGTTTACCCCCTGTATGGACCGCTGGAGGTCCGGGGTTAAACTATCCACGAATAGTATGATGGCCAGGAAGGTGTTGAAAAAGGGTGTTCGATAAAGATACGCCCGAAAACCATGTTTGTGAAGTCAACCTACCATGGACTAATAAATCGTCTAGCCTAAACGAATCAGTCGATGATTTGAAAGATGGGTTCCCAGATCCGAGTGCAGTCACTTCAAATTGAAACGCCCTATTCCCGGCCGCGGTTGCACAACACAATTCGCTAAAGAGGAAGACCCCGGCTCGTCCGAAAAGTGAAAGTCCGACAAGTAGCACATGTCGCGATAGCAGGCGAGGGTGTATAAGAATTGCTGCCAAAATTGGAACTAACTGTCCTAACGTGGCAATTACACTGATCCGCCGAACCGTTCTCGAGAAATCGCACTTAAAGACAGATGATTTGTCTACGCCAAGGCGAGGCTTCGAGATGGGATCAACAAGGGAGCTACTGTTGTGGCCCCGATTGCGACGACAAAGAAAACTGCCGACGTTTGCCCACAGTTCCGCAAAAGGTCAGACTCCAGCCGTAACACGTAGACGCTCTAGCTTTAGAATGGCAGTTAGGTTGGTCTGCTGAGCCACTAAGAACGCATTCGTCTATGAAGCCAGCTTGGAGCATCGACCCCGTGTCGTAGGTGTCATGCATTAAGTGAAGCTGCCATGCTAGACGCGGTCAGTGTACGAGACGGTCGATGCTTCATTACATATTCGAGATAGCGTGTTTGATATAGATTACTGCACCTGGGCGGGCGGTCAGACTGCCTTCTGCTGAGAGCCTTCCAGCGACTGGAATATTCAAGCGTACGAGTGGACTCACGACCTATATCATCTTTATTTTCTACACCACCGGAGCCCGTCTCTGAACATCTCAGCAATAATTGCTCAGGCGCAGAGGCCGACTGTAGGTTATAGTGGCGCAAGGCTCAGACGTAACAGAGAACGCTTAATGAGTCGAAGCGGCATTCCATGACGCCGTTACCTTGCCATGGGATAGAGTCTTCTGATAGAGGGGACTTCAGACAACCGGGGATCTGTGAGCGCTCGGCAGTAGTCTGGCTTATAAAAATCCGTCCGTTGTGTGGGCTGCTACGTGGTCAGAGAAGCAGAGTTTTATCCTCCCGGCAAGGGCTATAACTTTTAGAAGGGGCCAGATTATGAGTATGTCATCTGAGCAGCCTTACCCAGGGGTCACGTAGGCCAGCTGTGCAGCTTACGAGTAAGTACCTCTATCAGCTTAAAAAAGCTCTTGGTAAAGGCGACTGTGAGCAGTCTATTAGAATAAAGCTAAAGTCTATTACCCGAACGCAACATTTTTCCGGTTTTGATAGGGCCGGTGCCCGCGTTATCTCTCAGCGGGTTTTCATTCGCCCAATGTGCCCGCGCCCTTGGCGATGGGACAATCTTCCACGTAGAGTGATCGATTCTTCGGATACACACAATAGTCTCGGAACAGGTGACTGTCACCGCATCAGCACTATCCAGATTCAAACCACGCAACTCTTGTCGATGTCGAAGCGGTACAATTAGTTTCGTTCCTGCAGATTGGGTGGAATCCTCATGCCCCGGCTACCCTTAAGGATTTAACCTGAAAGAAATCATAAAATGCCAAAAGACAAAGCGCAGACCGCTGTCAGCAGGCAGCATAGTCCCTATGGAGCACAAAGTTCTTCGACCACCACTCATCCGTAGCCATCTGAGCGGGGAGATGCCGAGGTCAGATGCTATGGACGGCCAGGGTTAGCAACCAGCGCCCTCCCTCAGTCGGTCTCCGACAATTCGGGGTATCCACCATAACAGTTTCTGTTTTCCGAGCAAGAAGTTACGTGGCCCGCCCTGTCTGACATTCGTGCACTGGATTGTGTATGGCACATTCGTCTCTTCACCCCGTATAGGTACGGCACGAGCTAGTCAAGGCTGCCCGTAGATCTCGTTATCCCGCCTAAAGTGACGGTGCGTCTTCGACATCGGCGACGGACCTGCCTCATAAGGTGATTTATTTTATACAACGGGACCCGCCGGTCAGGACGGTGAAGTCAAGGTTCTATCTAAACTTCTGCTGTCGCAGTTCGCGTCAAGCCGAGAGGCATTCGACAAAGAAAGGGATATGGCACGGTTAATTTCCGCGATCCTCACCTTTCGGGCCACCACCGACTGAACTTACTGCCAAGCCTACTGAGCCAATATCCTCTTAGCAAGCAAGGGATTTCGTCCTCAGCATTGTTACGGTGACACACAGACTAGGGTGGGGACCATACTACTGAAAGGTCTCAACGTTCCATTCCCATACCTATACGCTGCATGTTACTCATAAGCGGGTTAGACAGTCAGTGAAGTAATGAGAGTCGGGGGACGTCGCGTGGGCCTAGATCTCAAAGTGCAGACCCCATTACAGAGAGCAATGAATTATATCATATATTTCCGAAAACCCGCCAGGCGTCGTTCAACGTTCACGCGCGCAGGAGTGGTTGAATAAGGACAAACGATTTCTCGAAAGGCCATTGATTGCACGAACTAGAAACAAGTACGGTTTATGTTGCTTGTATGTTCATGCGGTAACGGGTTCTTTACTCAAGCCTAGCCTGACTAGCGAGGAAATTTCGCCAGGGTTGGCCCGCCGGACTACCATATCTTGTGCAGGGAAGTGTGGCTACCGGTAAGCTGTTCTGACAAATTGATCTCGATGTCACTGTAGGCTTCGGCATCGAAGCGGAGTATACAGAAATGAGCAGAAATTGAAACAATCCCTGTATGGATGAACTTACTCCTACGTTTAGAGCTGAGAAACGTACTAGGTGCTTAAGGCGTCTGAGATCCAGACAGTAGTCGGCATACATACCATCCTCACGAGTGTTTAGAAAAATCGGCCTTGTTTGAGTTGCCTCGACCTCACTATACGTATTCGTTCTGAACGTCGACCTGCATTGTGAGGTAAAATAAGCTCGGCTTGTATGTGACTGTAATAGACGGAGTACTAAACTGACTAAGCCATCGCAAAACCAGTTCTAGATCGCTGTGGTTTCGTCCCGTTATCTTGTGATTGAATTTGCGCCCCGTATAGTGGGCACAATGACTATCTGGCAGTCCAGTCGGGACAGGGTTCCTGAGCTGCTTGGATGTGCTCACGCCGGAGCCCTAGGAGGGTCTAACCTTTGTTATGTCCGCACCACCTATCAGACGGTTTTCACGCCCTAGGGGCTCTGGGCAACTAGGAGCGTAGGCTTTGCGTGGTATCGTCGTGCTCTCTTACGCCTCTTATGGTTTATACGGTATTTGCAAATGTCGCGCAGTCGATCTGCAGATCCCGTCATATTCGTGCATAGGAATACTGAACGACTGTTCGCCGGTCGGCGCAATTCTGGTGTAGCGCTCCTTAACTCATGTGCATTACACTACAAAGGGAACAGAAGCCTTTAGAAAAGCAAATGGCCTCTATCTCTTTTGTAAATTCTGGAAGGACGCACGGTCGCATTCCGGATGCAACCGCCCCCCTGAAAGATCTGAAAGTTATTTAGCTTCTTGACCATAGCCAGGAATTAAAAGAGGGAAACAACTGACGAAAACGAGAGTGCGCACTTTTGCGTCTTTCATAAATGCTCGGCTCACTCTCAGGTTCGTCTAGTGCCAGGCATGGATGGCGAGGGATTGTGACAGTTTGCGCCTTGCCCTCCGCACGAACATCCCATGCGAGTCCTATCTGAACGCGCTCGCAACATGAGTAAATGAAGTTGGACGTTCTAAAAAGGGTAGGCCCATTTGCAAATGTGAAGGTAGCGGGATACGTTTTAACTCATAGTATTCATTCGTTCGGTATTAACTAGGTCTGACGTACGAGTCCATAACGGCGATTGGGTTCCCAGGGGCTTCTAACTTATCGTCTTGGCCAGCGGCAGTTGATGACTACCCAATTGGTGAGAGAGGATTTACTTTACCGACGATCCCCAGGCACAGCGTCCGAGGCTATCCTAGCTATCCTGTAAACGTCATGGAAGATCGCCCGAAACCTACCCGCTCGTCAGTACGAGTTGGAAGCGATAACGTCACTCAATTGGAACAATACTGTGTCGAGACGAGGATGTGAACGCGTTACCAGCCGGATAGCGATCGAACGAGCAGGATCGTGGGGCAAGAACAAACGTAAACGTCAAGCCGAACTACAAATTGTTGGACTAAGTGAGCACGCTATACAAATTTGCCAACTTCGCAGCTGGCTAAGCGCGCGTCGATCCAAACAGGACCTACAGTCGCGCGCTGGCCTATCCCCAATAAAAAAAGATATTAGACAGCGAATGCGCCACCTCTACATCTCGGGCCAAGATACATGTTGGTCGCTCTCGCGTACCCCGTAAGTGGTATCACAGCCCCACCAGATCGAGACTCGCGAGGCATCAGCCTGGTTGCCCATAAGAGCATACCCCGTTTCTGTCCCGAACCTAATAGCAGGCTACAGACTTGGCGTCAGTGAAATTCGAACGGCGTGAATCATAAAGAGATTCCTATTATTACATGAAACTGGACTCACAGGCCTCAGAGATATTAGGCGGCTAGCTCATCCTCATTAGTAGCTGCTCGCCATTGGTTCAACGCTCTGACGTGGGCATCGTTATGACCTTAAGCCGATCCTTCCAGCCTCCGCTCACGCTAACGACTCAGAGGTAGCTGGGGACACGCGGCGGGTCGCTTTCCGTGGTAGTTACGAATTTGCCGATGTGTCATACGCGCATTACAGTGATTTGACCGACAGTAGCAGGAGGATGCTATGTGCCATTCAACGGTGCAGTAGGCAAAGAACTTGGACTGACATAATACGTTCGTCCGAAGAATCGTGTTACTTCTAGTCAACAATAGATACCTGAAAGTATGATCACCTTCAATCGAAAAAGCCAGTAATGCCGGGTCGTAATACATGCAGGTCATTGACGTCTGCATTTCTCTTGAGCCCTTCATTCTTCGCCTTAGAAGATTGCACAATTTTTCAGCGGGGCACATACCTGCTTGGTCTCAATCGAATCGAATAGTAGAGACGCCACGCATTATGTGTCTCTTGTCTGATCCAAGTTAAGGCTGTTTATAGGGCTTGTCTTCAGATTGCCGCGCGAGCCATAGACTCGACAGACCACTATGAACGAATACCGGGTTCCAACTTACCGCTAAACAAAGGGTTACGCTGGAGGCCGACGCCCTTCCTCCCGAGGAGAACTGTCTGGCTCTCTATCAATGTAGATTAGATAATAGATGTATGACATTGACTCACTTATTGCGGTTGCTTCCATAACGAGAAAGGGTGTCTAGTGGCCCCAGCCTCGTACAAATATTCGCCCAGTTCTCAAATGTGGTTACGCCTAGGCTCCCAGAAAGCTTTTCCAAACCGAAGTCATAACACTTCTCAAGGGCTACTAAAGTGAATCGCAGGAGGAATTGCGCGGAACTACACTCCGGGATACGCCCAGAACAAGTAGGGGCGCGTCATTTCAGCTCAGCTACTCGTCACTTATCCTTAGTGCTCAGAAACTTAGCTGAGCACTAAGGGAATTTGTCTTCTAATAATACACTCCTGACGCCTGGCGGAGCGTTCCCTTTTTAAACTACGAACGCGCTTATCACATACCTGTTATTTAATTGACGTGTCGTGAAAGACACGTCCTCTTGCTGTATTTTACGAGCTAACGCAGCGGCATGGCGTGTCAGTGTGTCATGTGTGAAACCTCTCGCGTAAGCCGTGCGCGTCTCGCTTCCTCAACGGACGTGCAACTACTCCCGTGTAAGTGATCTCTTCCCACTACAACACATTAGGAGATTGCCCGGTGTACACTCGAAAAAGTCCAACATGAAGCAGGTGGGCACAAACTCAATTCGCATTTGACATTCATCGCTTCGGTCCACGGCTAATCCAGCGCGCCGTTGTTGGAATCCCGAAATGCGACTTGTATTCAAAGCTCACCTCTATTCGCATACACAGAACAAGTGCTGTAGGAAGACCTGGAGACGATCACGGCATGTCTACAACCCCTGCAGTGTAGAGAGCCGTTAAACTCGGGTGTAGTTCTGGTTCGGGCCGATGTGTTCTCGCTACCGATCATATTAATTTTACTAGAAGGCTCCGGGATGGGAGGCTTGCGAGGCCGATAATCGCTCTGCCAATTGCCAAGAGAGGCTATATTCTAGATGGGCTTATTCCTTTCTGAATGATAGCATTAACCCACCGCGGAGGGACGAGTGTCGCAGTCGGCAACGCCCAGGTGGAGGGTTCAACTAGCTCCGACCACGGCAACGCGTGCAGGCACGCCTCCGCCTCGGTTCCCCGGCTCCGACAAGTAGCTGCAAGTAATACATCAATTTGCCGAACTCTCTGGCGTCCTCGTCAGAATGACCCTTCAATTGTCACAAGGTCCGGTGTTTAGAGAACAGTTTGCTATCTGGGATCATAGAGATAGACTACCGGTAATAACTTTGTGCCACAATGTAGCTGACAGTACGGATCAGGCAGTCCTATATCTTTTCCGTTCCCGCAGCCCTTCGAGTAACTCGCGAGATCTAGAAGCTAAGGTCCAAACTCAGGATAGTACCGATTTAAACGGTACACCTAATATTGGACCCTCTCTAGATTTCAATGCATTTGATGTGAGTAGCCTGCTGCCGGATCCGCCGCGATTGTAGATGTGACTTCCTTAATGGTTAGTGACTGAGTGATATATGGGATGTTATGGAGTTTTTTTGTACTCGCACTACTAGTCGCCGTCTTTCATCTTGAATCAACTTTAGCCGGGACTTATGGGAGTGAATAGACCCGGCAGGCGGTGTGCTAATGTTATCTTGATAGCACTGGGTCATGACACGAACAATGCCGGCAATACTTTCAAAGAGTAACCTGGAATTTAAATTTGTCACCCGCGAACAAACCGGCCTAGAGTACATGGTCACAATACCATCGTAGCGCCGCGCAAAGAGGCCTCTTCTGTACCAGCTACGTGGAAGTGGGCGCAGGTCTGGTCGCTAGCGGCGTTGAACGAGCGAGCAAGGCCTGACACGCTAGCTTCTGCATGCGAGGCGTCTTATAAGCTGGCACCCCCTATATGTTGGGGAGCATCCACCCTTCTCAACACATGTTCTGATAAACACCCCTTTGACGTGGTCCTAGCGCGACTGACGCATAGTTCAGGGTGACTGAGATCTTCGGTGTGGTCTCCCCACCACTCAGATTGTTAACAAGCTGGGGTCGCCGTGTTTAGTTTTCATTCGTCGACATACCGCGGTGATCTCAGTAACTCATGAAGCTTTTAGTGACTACGGCGGCGAGTGTTAAAACCTAATTAGTCGCAGTTTCATTTAACGTAGTACCTAGAGAGCGATACACTCTTCCTCCCCGTAGAGAGATTCTCACGAAAAGTTCTGCGATTTGGCGTGTGACTCTCTTGCTCGAGGAAGTCTCTCAGAACATCCTTGAAATACTCAAGCTAAGTCATATACCACGCACTGAGGGAGGAAATCTTTGCTAACGAAGGGGATGCCGTTCTTTGCGGCGAGCAAGGGACGCAGATGTACATCTAGAACTGGATGGCGCTCGTAGAAGAGGGATAGTTTCCCGCAACCTCTGGGAATAAGGTAGTGTTCCCGTTTCCGACGCCCGCGACGAATCTACTAAGTATCTTGCTCGGCGATCCATCAAATTCTGGCGCGGCCCGGGATTTGTGGCTCCATTCGTCGTACTAATTCTTGTCCCATCGACGTTGGAGGCGCGCATTACTTTACAGCAAGCCGTAAGCAGACATACTGATACTCTGCGCTATTGGAACGCTAGTCACCCCATCTAAGACCTACGTCACTCCTGAATGCGGCTCCCCTTCTCGTCAGCGGACTCTCCACTCTACTCACCCCATACAACTCCAACTTTAGGGGCACGAGACTTTTCATGGTTCTAAGCGAAGATACTTTTATCTGACGCAAGCTCAGAGCCCTTGTGACCGTTAGATTGGCAGAATTCGAGTCCCAACTCTTCTGGATGGTAATCGGCTAGACACTTAAATGTACAGTAGACCACAGGTAGCATATGAGAAGGAACTCACGTACCATGGTGGGAGCTCCCAATAAGATGGGCTGCTATACTGCAATGAGACCTCCACATCAGGCGACGCCATACCAGTCGCGGATACTTGCTTCACCTTCAAATTGTACGAGGTCCCGCAGCAGGAAACTACGGCCTTTGTGGGTCCTTCGACTCAACGAAAGTACAGGGTACCTCACACTCTAGTTATCCTGAACTTTACAGTGAGCCACTGGGTGGCCTGCCAATATTCTTAAGTAAATATGGTCGTAAAGATTGACCTAACAACGAAATGGGTAACAGTCACCCGATGGCATCCGCCAGCAATACTTTTAAGGAGACATGAACGCCTTGTGGGTAACCCTGAATTGAGTCGTAGCAGCCCCCGACTACGCGCAGAACAATTCAATCCGCGCGTAACATTCTGCTAGACGGTTTGAGGTTGTGAACGGACTTTGGGAATATATGATCGTTCGACTGGTCGACGCACTGTTCAACTAGACTCAATCCGGTGTGGGAATAAACACCACATCTGCCCTACATCGACACTTTCCGCGCGTTCGACGGTTCGGTGCTACGACCTTTTTGTGATACCCTTTGGGCCTTGTTGCACGAATGAACTACCATGCAAGTAGTCTACGCGTCGAGTTCCACGTCAGGCCGGCGCCATTCCAAAGCTTATACAGATTACCTTGACTCAGGTCGACGCTTGGGACCAAGTGGCGTTGCCTCTGGTCGCACATAGATGCACATTCTCTACAGGGTAATTGAAGCTTGTAAAGTGCGTAATGAAAGCTATTTATATGTGGCTGCGTCGTCGGACTGCCGACATGCGCTCCCACAGGAGGTGCATTTTTTTGACTCAACCCTCGAGCTTAGCTTGGAAATGGCAAGCCCGCCTTATAAGCTTCATATGCTCGGAGGACGGTCCCTGAGCACGAACCAGTCGCAGCACGCTCATGATAGGTTTTCTGGTTCGGTCGGGGGCGCTCAGGATTGGTTCTTGATGGCTCTTGATGTCATAATTTTAATTTGCCAATTATTACCCCTTTGCGAAAAAATCACATCCGCCACGTCCTTCTACCACTGCCTTCTATCAGCAAAGTAGTCAGTTACCCAGTCCCGCACGTGACAACCCTCTAGATGTAATGAGCGTGCCGCCATTGTGCGCGAGATTGTAGAGGCTAGTCAATATTGGGGGCCAGTATTCACACTGATTTGAGGACAGTTAACCGATCCTATAATGGTCCTGCCCCACAGCGCCCCGAATTAGTTGGAGGTAGCACAGGGCTTAAATCGAACGCAGCAAAGGGGCCAATGTCAGTCTGAGAGTGTCCCGGATAAGTCGGATTACCCTTACGAGGGTGCTAGAACCCCCGGGAGTGGTCCGAAGATCAGGGCGTTCGGGTTACGTCTTGTGATTCTGTACGGACTCACCCTGACTCATAAATGAAGCGAGCAACCAATCACGCTTGCGCTCCTCTCCATGGACCTCTTGCTTCGGGGACGAAACCTCCCCACTTCCACGCTCCTTCGGCGCCAAGTGTACATGAGGCCTATACGAACGCACGGATTGTATCAACGGTTGTCAGTTTGCGGTTGCAATAAGTCCAACCTCCAACGTGTTCTTGGCCCCATTGTGTGTCAGTCACATCCCGAGCCACAGCGGCCGGTCTTGGTCACACTCTCGTATTTTGAGATGAAGTTGCTAGAGTGCCTCGAAGTCCCAGTTTTCGTTGCAGAGTCTGTGGTGCTTCGGGGTCCGCACCCCCCCGCCCGGGCCCAGGATGGGAGCCGCGTCCTCGGAGATGGGACCCACCCACCGTCGTTTTTAGAAGCTCACAAATCGGCTACATGCACGGCTATGCAACTAAGCCCGTCGTTCCCTCTGATCGACGTCCCGATCCTCCGGACTCGCACTTAACGCTTTCATATGAGCCAAAAACAGACGAGGCGGCGGCATAAGAGTGTAGGAATCCGTTTCGATCCGACGCAGGGCCGTAAGGGCACGATTCATTCCTAATGGAAATACGCACTCAATTCTGAGCAGCAAATATGATTAAAAAAGTGTCCGGCGCCCCTTATCCAAGGGAAACTTCGACGATATGTGGGGTAGACTTCCCATCGGCGTAGTACCAGATCGTGCTACACTCTCTCAAGGTTCCCAAACTATCGATTACGATGCTTCGAAAATTTTCTCAGCCTTGGCTAGACAATGGTTGGAATCCAAGAGCTAGTGTGCCGCCACGTGATGTTGTGAACCCTGTCTAGTACTCTAATGATTAGGGTATGTGACCTCTCAGTATGTGCCATTTAGGGACACCAAGCGGTCTGGGTCAATTCCAGTGGGGAATACTATCTAGGAACCTGAGCAAGACGAGGGAAGTCTAAGGAGGCAAGTATCCTGTGCGCGCTTTAGTCCGAAATTGGAGCAGAAACAAGTGGTTGGATAGCCAAAGCAAGGTGACCTCCCCGCTCGTCATCGCAACAAAGGCGGTATGACCCGCCCATCATGTGCCCGGAAACAATAATCTGTGTGTTGGGTCTGGAAAAGACATTTACGAATTCCATGGCAATAGTATGCAAATGGCCACTCTTGACGCATTCGTTATGTTGGACGACGTACTGCAGTCTGGTACTAGTCCATAATACATAGGGACACTACACAGGTGTCGACAAGACTAGTGGAGTCGTACGATCGCCCCCGCCTCTGATGGCGATATATCGCACGTGTGGTAGACTATCCACCGACATATTTGGTCCACATACAATGCTCCGCTGCGAGCATACCGATCGGATGTAACGATTGTGCCCAAGTCCGCGTACTAGCATCCGTCCTACAGGCCGCGAGATGTTTAATCCGAGTAGAGCTGAGGGTTAACTATGAATTTTGCAATGGCTTAGCACACCAGATGTTGAGGGGTATTAGTTTCGTACTGTTGTACTTACTATCGCAGGGCCCACCCACGCCCACTTACCTTCCAAGACAAGGTGAGAAATAGCAATGGGCAGGGTGATGCTCTGTGCTGAAGGACGGACAACGTACGTCGAACGGCGTTGTGTTCGACTAGAGTGCAGGGATAGTAGGCTGATCGCCGACACATTAAGGCAGACGAACCCAATAACAAGGCAAGGCTCCTCAGTAAGCACCGCGTCTGAGGCGGCATTCGGCTCAGCGACCCTTGGAATAGGCGCCAATAGCCTAAGGCCGCCAACCAGGATATTCGCTCTCAACCAACCGAGTATCACGCGAGCAGGAGCAGTCTATTATATAGGCAGGGTCCTGCGGGATTTGCTCTCGGGCTAGGGAAGAACGAAACTAGGATGAGTCGACGCGGCGACGGGTGCTGAGTTTCGTGGAAGCTTATGCTTCTCATCGTCGGGAAGTATACATATGCTTCAATCATCTCCTTGGATACTATTCTACTAGGCGATACCGACTTAACAGAGCGGCACATTCGGCTAGGCAACATGTCACATATGACAGGCACCAGCGAGAGCATCAGCGACCTCCCATTAAACACTACGTGTAGAGTCAGCATCTAAACATTATAAACCTTCGTGGAACTTTAATCCGTGCTCTGGTAGCTGTCACCGTTCTAACTCTCGCGAATAAAAGCGCTGCCCATAGCTTGTTGTACTGCCTACAGTGGTAGATGGAGAGCATTACACAGTTACATGCCTCTGCTGGTATGCCCTATACACGTCCTGAGTTCACCACACAGTAGAGTCTGCGTTATCTGCAATCCGGGCGATGATGAAGCCAGAGAACGTCCGGTAGCGCAGTAGGCCTCGGCGCTCAGGCAAAGCAATGAGCCGGCTGGGAATGTCTTTGCGGCAAGCAGAAGGTACTCGCCCATATACGGACTCACGTTAGTACCCCGAGACGTAATAAAATCAACATTCGACACGTTGGACCGTATCTGGTACGAGCTCGATAGAGCTTTAGAGGGGCGTACGCGATTGAATACTATGTCCTTATAGTCGGATTGCATAACCTAATGTCGGGGTAAAAGAGCCTACATTAACTGGAGTACCGATCTCAGATGGTGCCCTGTTGCATCTGTAATCCGTCCAGGTGCTTGAAGGTACGCAAAAGGTGCTAGCTGATTGATGAGTAAATAATGCATACCATCATGGCCAAGACTTGTGATGGCCAAATTCTTTTGCCCACGTCCAAGTAAATAGAACCACATTGCCCGAAGTGGGGATGAGGCGAAACTCGCATTTGCCAAAAGAGTACAGCTTTTAACCTCACCCCATGGGCGGTCGTACAAGAAGGGAAACGTGTTCCTAGTGACGAATTAAGCAGGTTTCCTACTCTTATAAGGCTGTTAATAAGTTGTGGAAGCGAGGCAAGACCTGGGATCTTCCTGAATTTGGGTTAGTTATGGCCCAGAGCTGGCTCTTACTCTTCATGTGGAAACTGGATGGGTAAGAGTCCATGATGAGTAATTTGTGGCGTGCGGGCTCAAATGTCCATAGTATATTTCTCGTGCTGAACGGACTGTGCGCACTACTGCCTGCACAAGTGGCACGTTTTTCACACGTGGGACGACTCAGCTATAAGTTCTACCGCTTGAGATCGCATTTACATATCTGCTACCTATATGCTAGAAGGAACCACGAGTTCCAGCAAGGAATGATTCATATGACTTCATGGTTGTAAGATATAGACGCCTCAACGATTTTCTCAAATGAGCGCTCGCCTAACCTAGCATTGGATTGCATGAGCCCATCGGCGCTCAACTCAGGTTTGACGCGGGATAATTAACAACCACATCAAGCCGGGCCCCGATTGGACCGCACGCCGCCCATCTCGTCTTATCAGGGCCGAACTCGTATTTGCGCTGTTCTCCTACACAGATCATGACCCAGACTCGAAAACCGCTCGGCTTCGGTGAGACGGGCACCCATAGCTCAAGGTTTGGTCTTAGATTAAGCCGATCTTCCGCATCGCCTGCCCATGAGGATCACACGCTCTTATTCTTCTTTGGACTGCGCCGTGTATACCCCGTTCAAGCCGAGAACTACATTTCACTGATAGTTTGTGGCGACTTACTCCCTCGCTCGTTCTGCCCAGTGGAAACTGGCCAAAGGCCTGAGGCACAACTACCTAAGACGACAGCCTGCTATTCGGCTGTATTCCTTTTGGGCCCCCACACTCTGGAAGGCACTACAAGATATTCACCTTCCCTAATAACTGGGGGATCGTCGTGCTGTGGAACTCTCTAGCACCGCGACATCCGCGAGTTGCGGGCCCCGTCATCTGGAGGTTTCTCTCGTGTCCTTCCAGGGTAGCCATCGGCGCGTTAGTATGTATATTTCGCGGTAATACTACTGGCGCCGAATCTGACCCCGCACAGGTTATAGCAGCGAAACATACCCTGTGGTGTACCCACACGAAGATGACCATCTAGAGGAGGAGGAAACTGCCAAGATACGACTTTGCGCTTCAAAGGGTGGCCACACGGGCAACCAGCACTTGCGCGGAAACGAGTTCAACAATACATGTAGAACCTCAAAACGTCGGATACACTTTTATGCCGCTGAGCTCGTGGGAGCGGCCCTGTGAGTGCGCTATTTCACATCTTCTCCTGACGTGAGTTACCACTGTCACTGGTAAGCACAGACGTTTAGAACATTTTCGTCCCAACCGTAACCGGTGCCGGTTGCAGCGGGCGGTAGGTGGGCTGCTATAGCTGTGGTTGAAGCCTGGCGCGAAGTTTGACTATAAAGAGGGTGATGCAATTGTGCTATTCATGCTATCTAGACGTCTGTCTGGTGATCAGGCGGGCGTAGGTCTCGGGTTACGTACTGCCCCTGTCCGCCTGAATCGCACACGTTTGTTAAAGCGCTTCGGTGCCTGGTTTTCAACGCAGTCGACGTAACAACTTATTCTTGTTTCGATATCCGATTGTCATACCGACGCCTCGAGTCTCGAGCAATCTCAGCGCCACTTGTCTTCCAGAGGAGCCTTTCTTTTGCCCAGTATCTTTCCTTTTTACTCACTGTGATAGTGAGCTGCATGCCGTTTCATCTGCTTGTGGCTGGGCTGCAGAACAGTTAGCTAGTAGTTTGTTGTACAGAGTTTTCGCCGCGCTACAACGCGCGTATGGAGACCGTTGTCATCCCTATGATAACAGTATATATTCCGGTAGTCGGAACCGTCACAGAGGGCTAAAGATAACTTCGTACTGATCACGCGCATTTCCTACGGGAGATGCGTGTCCAGGCGTTAGGCCGGGGAGTGCCAGGGCTGCATAGCGTACTCCTCCGCCCATCGAGTACCATGACCAGTGAGCACCGGGTGTGTCGATTGTTTGTTGCGGATGACCTTGAGAAAGACGACCGTCGGGGCGGCCTCAGAAAAGCCGGTACAGACTTCAGCTAACCTCGGGGTACATACTTACGGTGAAGTGGAGAGAGATCGGATCCGCTGAGTCGGTGTGAAAATAGAATTGTGTGGTTCTGATACGAGGTTTTATTGCGAGCTGTTTTTGTGCCATATGAGCCACCTGGTTCCACCTGGTGGGTCACGTGGGGGGAGGCGTATAGTCCGGAAGATTTCTCTCTATAGCACCGATGATGGATGCACCAGCACTCCCTGTCGGTTTTCTACTGGCCCCTGCGCGCTATCGCGCCCGCAGACATTGACAAAAAGAAAGGAATATCTCCACATTGGCATTACCGATTACCCTGGGCGTGCGTATGTTTTAAAGGCTCCGTACACCATACCTAATCTCAAGATCCCGGGGGCTCGCCCGCCAGATTGTTCGATCTTCATTTATTCCAAAGGCTCGTCTCGAAAGGTCGGGTCTGTTAGGTTCGAGGACTATACATGGAATCGAATGCATAGACGGTTTAAAGGGGGCTATTCTCATGAGAGCCTGATATGGGGGGATTGTGCAAGCACTTGAAGGCTGCAGCATAACTAGGTGCCGATTCATGTCTATGGGTGACCCACTTCACATCAGCGGTGTCACGCGGGGATGCGCGCAATTGTATGTCACCTCGGGGTCGACGTTCGTAGGCCAATTAAGAGCCGATCTTATCTTGAGCGTTTAAACGGTGGCGAATTGAGATACGTGTCCCAGACCGGGCTTGTGAACCTACCCGAATCACAGCGGGAGGACCCAGGTCATGCCCCGGAAGCAAATACTTGTCACCGCTACCACCTTTCTGTGCCCTTTCGAACGGCTCTAGTGGTACGCTAGCCTAAGGAGGACCCTGGCCCCCAGGCTTGCTAGTTTAGATAATCAATGCATCATATTCCAGATCAACTGATTTCGCAAACACGTTTGACGATTACCGGGCCCTCCCGGGTGAGGAGTTCGTCCGTCCCCGGGGTATATTCTCCTCCCCCTGAGCTGTCAGCTTTCGACTAAGTGGACCGCCCGGATTAAACGACATGGTTATTGAGGGACTCTGAAACCAGGGACCAAGAGGGCCGGTTAGATTGGTCCACTCCTAATCGGCTTTTCTTGGTTGATTCGTCTTTTTGGGAGTGCTTGATGGCGTTTGGGGGGCTAATTCTTTCCACTATGTGAGATGAGACTGTTACGATCTTTTAACCAGTACTACTAACGAGTTGAACGATGCGTTAATGACGAGGCGGTATTACGTGACACCCGGTATCCGAGCAGTACTCTTTGTGTCTGGACAGGCGTGCCATGGTACGTGGATCGTCGTCGCGACTGAGCCAGGGATTATCGAGCCCTCTTCCCTGGTCAGCACGCTACGATCGTATAATACCTAAGCCGCGGTAAGAGGCGTCGGGGAGGCTATAGGAATGATAAAACGTAGTTGTGCATACCCTGGTAGCCTGAAGCAGGACGGCTCGAAGATCACGTCGATGGGCTATGATAAGGATAGGTGGATGAACAAACAGAGTTAGGCTATATGTAGGTAAAACCGGATAAAGGAGACACGGCTATGACTCCACGTAGCCGAAATGGGTGGAAGAGTTGTGCCCCATTAGTTCATCGTCAGTAAGCACGCCGCTAGGCTCGCACCAGAGTGCCTATCGATCATGAGAGTGAGGGCTACTTTCTAAATCACGGAAGTATGAGGAACATGCCTTTGGCAGAGCCTGTCAGTATTCGTTGGGTGACAAGCACGACCAGTGCCACTCGGAGAAGCCGGGCCGCTAACTCCGGGACGACGCTAAGTATGAACTACAGTTATCAGCTGTACGCTGTATAAGATTAAGGGTGGCATGCAGGACTTTATGATTGTAGCACATTGAAATGAAGTTAGGTGACTCAGCTACTCCCAGAAAAAGCAACAAACACTGCAACATGATTACGCGCGTTTTGAACATAAACTTGGGGGGGACTCACCCTGAATTTCTAATGCAAATACAATAATTCGAGCATCGTCCTAGATGGTCCCTCGCCAATTGGATGAGGTGTCCCACATTGTCGGCACGCGTACAAACCACGGACTTATCGTAAAGTACGCAGAACCACATGTCTGTATTAACTGGTCAACGCAAATACTGAGACGTCTATTTTTCGTGCTCAACTTCCCCTAGTCTGGATCGGTGCTGCGCCGCTACACTCCGTTGAGTGCTGTAGTCTCTTACTTGACACGCTTTTTGCAGAGCCACGGGCTTTGTACACTCGAGCCGGAGCTCAGGTTCCTCGGTCACGACCCCTAACCAACTACAACTCCGAATTCTGTGCACGTCGTATTACTGGATCTTCTAGAGTGAGGGTTCTACTAACGCTACCAAGCGCCAAATTGGAGCGAATACCATTCCATACCGGAATTGCCCGCCGCCCTTCAGGGGAAATCTTCCAGTCACATCGCTGGTGCTAGCTGTGCCACGTTGGGTTCCTTGTAAGAACTTCTGTTACTTGATCCGCCTCTTAAAGTCGGTTACCCAGGCAGGGATTCCGAGACCCCCTTACATGCGCTATCTACGGGGACGCTGTCTATCGTATGAGTGTTTGACTTGTAATTCCCGGACACGGAGGCGTTATCTCGCAGCCAGTGGCGCGAACGCAGCACCCGGACAGCAGTCTCCGAAGCACCATCCGCATCTAAAGTAACTCGGGCATCCGCGGGTGCAGCGTTTGGAGATTGTATGCCACGTTCTTCGGAGTTTTCAGTGTGGCACGACCGTGGATGGATTAATATCCCGCCATCTCGGCGAGTCCAGCCTTGCGTAGTCCGGCTTTACATTATCCGCCCAAAGCCAAATCTCTCAAGAAGGTGCTTGAAATGCTGAGAGAATGCGACCAGCGCCATGGGCGATCAAGATTGTAGCAAATTAACCTAAGCCGTCGCTACAGACTCTAAGTTAATAAAGACGGCCATGGGTCATTTTTCGAAATACTTCGGTCTGTGGTACACAGCTTTTAGTAGAGGCCCATGTCCGCATGTTAGCCTGCCATAAGTGCCCCTCAACTAACTCTGAAGCGTATGTGTGGTTAACTGCGGTGCTTCTAAGCATCCGGTCGGAGGGCCCGCGTATTTTTATCAGCGAGAATCTAAAGGCCGCACGATTCGGAGTCTAGAGAACCGCCACGCGGAGCCTTCGGCGACAGTTTTCTTCGTTACGTAGCTAGTCCACGGAGCCTAAGCTGCCACATGCGACAGTCGGACTCTAGTTGATTGTGTTAGGTGAAAATAAAAACTTGGACAAACGCCTCGGAATGCACGTTCTAATCCTCAACACCGAGGGGCTCCCCCATTGCTCTGAACTTATACCCATCTGATCTAGTCGGTGTGCGGAGTGTTGGGACCATACAATACATAAGCTCTTATTAGGCGCAGATCCTGGAACAGCAAACTGTCACAAGATTGCGCTGGGCACCGAGCTCTTAGTGAGGCGAACCGTTGGCAAGCTTAGCGAACACTTCCGTCGTAGACTGGTCCTCACATCGGTAACAGTACGAATGCGTTCAGAGCAGTGTTACGCAACCGCTTTGCCAA";
